requirements.txt
flask
transformers
torch
Pillow
gtts
flask-cors
